// models/enums/skill.enum.ts
export enum SkillEnum {
    ACROBATICS = 'Acrobatics',
    ARCANA = 'Arcana',
    ATHLETICS = 'Athletics',
    CRAFTING = 'Crafting',
    DECEPTION = 'Deception',
    DIPLOMACY = 'Diplomacy',
    INTIMIDATION = 'Intimidation',
    MEDICINE = 'Medicine',
    LORE = 'Lore',
    NATURE = 'Nature',
    OCCULTISM = 'Occultism',
    PERFORMANCE = 'Performance',
    RELIGION = 'Religion',
    SOCIETY = 'Society',
    STEALTH = 'Stealth',
    SURVIVAL = 'Survival',
    THIEVERY = 'Thievery'
}